DAG for funcargs and array elems Inline optimization WIP VarUse command in LIR Instruction planner with linking variable uses with arrays memory selection in array declaration symtab varcopy without memory and value copy

+++ include/symtab/vartb.h
@@ -39,8 +39,8 @@ int VRTB_update_memory(long id, long offset, long size, char reg, vartab_ctx_t* 
 int VRTB_update_definition(long id, long definition, vartab_ctx_t* ctx);
 int VRTB_get_info_id(long id, variable_info_t* info, vartab_ctx_t* ctx);
 int VRTB_get_info(const char* vname, short scope, variable_info_t* info, vartab_ctx_t* ctx);
-int VRTB_add_copy(variable_info_t* src, vartab_ctx_t* ctx);
-int VRTB_add_info(const char* name, token_type_t type, short scope, token_flags_t* flags, vartab_ctx_t* ctx);
+long VRTB_add_copy(variable_info_t* src, vartab_ctx_t* ctx);
+long VRTB_add_info(const char* name, token_type_t type, short scope, token_flags_t* flags, vartab_ctx_t* ctx);
 int VRTB_unload(vartab_ctx_t* ctx);
 
 #endif
